fastapi-backend/requirements.txt
FastAPI
uvicorn
openai
python-dotenv
